修正1

+++ app/controllers/items_controller.rb
@@ -38,8 +38,11 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item.user != current_user
+      redirect_to root_path
+    else
+      item.destroy
+    end
   end
 
   private
